+++ package.json
@@ -1,42 +1,45 @@
 {
-  "name": "gelato-instadapp-debt-bridge",
-  "version": "1.0.0",
-  "description": "The smart contract and tests for gelato-instadapp debt bridge",
-  "repository": "https://github.com/joelamouche/gelato-instadapp-debt-bridge/",
-  "author": "Antoine Estienne",
-  "main": "index.ts",
-  "private": false,
-  "scripts": {
-    "build": "npm run compile && npx buidler typechain",
-    "compile": "npx buidler compile",
-    "test": "npx buidler test"
-  },
-  "devDependencies": {
-    "@gelatonetwork/core": "^0.6.0",
-    "@nomiclabs/buidler": "^1.4.7",
-    "@nomiclabs/buidler-ethers": "^2.0.0",
-    "@nomiclabs/buidler-ganache": "^1.3.3",
-    "@nomiclabs/buidler-waffle": "^2.1.0",
-    "buidler-typechain": "^0.1.1",
-    "typechain": "^2.0.0",
-    "@typechain/ethers-v4": "^1.0.1",
-    "@typechain/truffle-v4": "^2.0.3",
-    "@typechain/truffle-v5": "^2.0.2",
-    "@typechain/web3-v1": "^1.0.0",
-    "@types/chai": "^4.2.12",
-    "@types/mocha": "^8.0.3",
-    "@types/node": "^14.11.2",
-    "chai": "^4.2.0",
-    "dotenv": "8.2.0",
-    "ethereum-waffle": "3.0.2",
-    "ethers": "^5.0.14",
-    "prettier": "2.0.5",
-    "ts-node": "^9.0.0",
-    "typescript": "^4.0.3",
-    "ts-generator": "0.0.8"
-  },
-  "dependencies": {
-    "dsa-sdk": "^1.4.40",
-    "web3": "^1.3.0"
-  }
+    "name": "gelato-instadapp-debt-bridge",
+    "version": "1.0.0",
+    "description": "The smart contract and tests for gelato-instadapp debt bridge",
+    "repository": "https://github.com/joelamouche/gelato-instadapp-debt-bridge/",
+    "author": "Antoine Estienne",
+    "main": "dist/index.js",
+    "types": "dist/index.d.ts",
+    "private": false,
+    "scripts": {
+        "build": "npm run compile && npx buidler typechain",
+        "compile": "npx buidler compile",
+        "test": "npx buidler test",
+        "compileJs": "rm -rf dist && tsc",
+        "cleanInstall": "rm -rf node_modules && rm package-lock.json && npm install"
+    },
+    "devDependencies": {
+        "@nomiclabs/buidler": "^1.4.7",
+        "@nomiclabs/buidler-ethers": "^2.0.0",
+        "@nomiclabs/buidler-ganache": "^1.3.3",
+        "@nomiclabs/buidler-waffle": "^2.1.0",
+        "buidler-typechain": "^0.1.1",
+        "typechain": "^2.0.0",
+        "@typechain/ethers-v4": "^1.0.1",
+        "@typechain/truffle-v4": "^2.0.3",
+        "@typechain/truffle-v5": "^2.0.2",
+        "@typechain/web3-v1": "^1.0.0",
+        "@types/chai": "^4.2.12",
+        "@types/mocha": "^8.0.3",
+        "@types/node": "^14.11.2",
+        "chai": "^4.2.0",
+        "dotenv": "8.2.0",
+        "ethereum-waffle": "3.0.2",
+        "prettier": "2.0.5",
+        "ts-node": "^9.0.0",
+        "typescript": "^4.0.3",
+        "ts-generator": "0.0.8"
+    },
+    "dependencies": {
+        "@gelatonetwork/core": "^0.6.0",
+        "dsa-sdk": "^1.4.40",
+        "web3": "^1.3.0",
+        "ethers": "^5.0.14"
+    }
 }
