+++ tsconfig.json
@@ -1,39 +1,33 @@
 {
-  "compilerOptions": {
-    "target": "es5",
-    "module": "commonjs",
-    "lib": [
-      "esnext"
-    ],
-    "experimentalDecorators": true,
-    "downlevelIteration": true,
-    "resolveJsonModule": true,
-    "esModuleInterop": true,
-    "noImplicitReturns": true,
-    "noImplicitAny": false,
-    "emitDecoratorMetadata": true,
-    "declaration": true,
-    "pretty": true,
-    "skipLibCheck": true,
-    "sourceMap": true,
-    "noEmitOnError": true,
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "allowJs": true,
-    "allowSyntheticDefaultImports": true,
-    "moduleResolution": "node",
-    //"isolatedModules": true,
-    "noEmit": true,
-    "jsx": "react"
-  },
-  // files:[
-  //   "node_modules/@nomadiclabs/buidler-ethrs/src/type-extensions.d.ts"
-  // ],
-  "include": [
-    "lib",
-    "test"
-  ],
-  "exclude": [
-    "node_modules"
-  ],
+    "compilerOptions": {
+        "target": "es5",
+        "module": "commonjs",
+        "lib": ["esnext"],
+        "experimentalDecorators": true,
+        "downlevelIteration": true,
+        "resolveJsonModule": true,
+        "esModuleInterop": true,
+        "noImplicitReturns": true,
+        "noImplicitAny": false,
+        "emitDecoratorMetadata": true,
+        "declaration": true,
+        "pretty": true,
+        "skipLibCheck": true,
+        "sourceMap": true,
+        "noEmitOnError": true,
+        "strict": true,
+        "forceConsistentCasingInFileNames": true,
+        "allowJs": true,
+        "allowSyntheticDefaultImports": true,
+        "moduleResolution": "node",
+        //"isolatedModules": true,
+        "noEmit": false,
+        "jsx": "react",
+        "outDir": "./dist"
+    },
+    // files:[
+    //   "node_modules/@nomadiclabs/buidler-ethrs/src/type-extensions.d.ts"
+    // ],
+    "include": ["src"],
+    "exclude": ["node_modules", "src/test"]
 }
